Move useLocalStorage out to a separate file so it can be reused.
