+++ website/src/components/logbooks/VisitedStamps.jsx
@@ -1,4 +1,5 @@
-import React, { useEffect, useState, useMemo, useCallback } from "react";
+import React, { useEffect, useMemo } from "react";
+import { useLocalStorage } from "../../hooks/useLocalStorage.js";
 import { Link } from "react-router-dom";
 
 import { pageMap } from "./LogbookContent.js"
@@ -101,40 +102,10 @@ function VisitedStamps({ currentPage }) {
     );
 } 
 
-function useLocalStorage(key, defaultValue) {
-    const [value, setValue] = useState(() => {
-        try {
-            const saved = localStorage.getItem(key);
-            console.log(`Loading ${key} from localStorage: ${saved}`);
-            if (saved !== null) {
-                return JSON.parse(saved);
-            }
-            return defaultValue;
-        } catch (e) {
-            console.error(e);
-            return defaultValue;
-        }
-    });
-
-    useEffect(() => {
-        if (value === null) {
-            return;
-        }
-        const rawValue = JSON.stringify(value);
-        localStorage.setItem(key, rawValue);
-    }, [key, value]);
-
-    const deleteValue = useCallback((): void => {
-        console.log(`Deleting ${key} from localStorage`);
-        localStorage.removeItem(key);
-        setValue(defaultValue);
-    }, [key, defaultValue]);
-
-    return [value, setValue, deleteValue];
-}
-
 function randomRange(min, max) {
     return Math.floor(Math.random() * (max - min)) + min;
 }
 
 export default VisitedStamps;
+
+
